+++ src/views/bbissues/headerNode.ts
@@ -4,14 +4,14 @@ import { Resources } from '../../resources';
 import { AbstractBaseNode } from '../nodes/abstractBaseNode';
 
 export class CreateBitbucketIssueNode extends AbstractBaseNode {
-
     getTreeItem(): TreeItem {
         let treeItem = new TreeItem('Create issue...', TreeItemCollapsibleState.None);
         treeItem.iconPath = Resources.icons.get('add');
 
         treeItem.command = {
             command: Commands.CreateBitbucketIssue,
-            title: 'Create Bitbucket issue'
+            title: 'Create Bitbucket issue',
+            arguments: ['explorerNode']
         };
 
         return treeItem;