+++ src/views/jira/headerNode.ts
@@ -4,14 +4,14 @@ import { Resources } from '../../resources';
 import { AbstractBaseNode } from '../nodes/abstractBaseNode';
 
 export class CreateJiraIssueNode extends AbstractBaseNode {
-
     getTreeItem(): TreeItem {
         let treeItem = new TreeItem('Create issue...', TreeItemCollapsibleState.None);
         treeItem.iconPath = Resources.icons.get('add');
 
         treeItem.command = {
             command: Commands.CreateIssue,
-            title: 'Create Jira issue'
+            title: 'Create Jira issue',
+            arguments: [undefined, 'explorerNode']
         };
 
         return treeItem;