+++ src/webviews/bitbucketIssueWebview.ts
@@ -1,20 +1,24 @@
-import * as vscode from "vscode";
-import { bbIssueCommentEvent, bbIssueTransitionedEvent, bbIssueUrlCopiedEvent } from "../analytics";
-import { DetailedSiteInfo, Product, ProductBitbucket } from "../atlclients/authInfo";
-import { clientForSite } from "../bitbucket/bbUtils";
-import { BitbucketIssue, User } from "../bitbucket/model";
-import { Commands } from "../commands";
-import { Container } from "../container";
-import { isCreateJiraIssueAction, isOpenStartWorkPageAction, isPostChange, isPostComment } from "../ipc/bitbucketIssueActions";
-import { BitbucketIssueMessageData } from "../ipc/bitbucketIssueMessaging";
+import * as vscode from 'vscode';
+import { bbIssueCommentEvent, bbIssueTransitionedEvent, bbIssueUrlCopiedEvent } from '../analytics';
+import { DetailedSiteInfo, Product, ProductBitbucket } from '../atlclients/authInfo';
+import { clientForSite } from '../bitbucket/bbUtils';
+import { BitbucketIssue, User } from '../bitbucket/model';
+import { Commands } from '../commands';
+import { Container } from '../container';
+import {
+    isAssignIssue,
+    isCreateJiraIssueAction,
+    isOpenStartWorkPageAction,
+    isPostChange,
+    isPostComment
+} from '../ipc/bitbucketIssueActions';
+import { BitbucketIssueMessageData } from '../ipc/bitbucketIssueMessaging';
 import { Action, onlineStatus } from '../ipc/messaging';
-import { isFetchUsers } from "../ipc/prActions";
-import { Logger } from "../logger";
-import { AbstractReactWebview, InitializingWebview } from "./abstractWebview";
-
+import { isFetchUsers } from '../ipc/prActions';
+import { Logger } from '../logger';
+import { AbstractReactWebview, InitializingWebview } from './abstractWebview';
 
 export class BitbucketIssueWebview extends AbstractReactWebview implements InitializingWebview<BitbucketIssue> {
-
     private _issue?: BitbucketIssue;
     private _participants: Map<string, User> = new Map();
 
@@ -23,11 +27,11 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
     }
 
     public get title(): string {
-        return "Bitbucket Issue ";
+        return 'Bitbucket Issue ';
     }
 
     public get id(): string {
-        return "bitbucketIssueScreen";
+        return 'bitbucketIssueScreen';
     }
 
     public get siteOrUndefined(): DetailedSiteInfo | undefined {
@@ -64,8 +68,9 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
             return;
         }
 
-
-        if (this._panel) { this._panel.title = `Bitbucket issue #${issue.data.id}`; }
+        if (this._panel) {
+            this._panel.title = `Bitbucket issue #${issue.data.id}`;
+        }
 
         if (!Container.onlineDetector.isOnline()) {
             this.postMessage(onlineStatus(false));
@@ -80,16 +85,17 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
                 Container.bitbucketContext.currentUser(issue.site),
                 bbApi.issues!.refetch(issue),
                 bbApi.issues!.getComments(issue),
-                bbApi.issues!.getChanges(issue)]
-            );
+                bbApi.issues!.getChanges(issue)
+            ]);
 
             this._participants.clear();
             comments.data.forEach(c => this._participants.set(c.user.accountId, c.user));
 
             //@ts-ignore
             // replace comment with change data which contains additional details
-            const updatedComments = comments.data.map(comment =>
-                changes.data.find(change => change.id! === comment.id!) || comment);
+            const updatedComments = comments.data.map(
+                comment => changes.data.find(change => change.id! === comment.id!) || comment
+            );
             const msg: BitbucketIssueMessageData = {
                 type: 'updateBitbucketIssue' as 'updateBitbucketIssue',
                 issue: issueLatest,
@@ -124,20 +130,23 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
                     const linkUrl = this._issue!.data.links!.html!.href!;
                     await vscode.env.clipboard.writeText(linkUrl);
                     vscode.window.showInformationMessage(`Copied issue link to clipboard - ${linkUrl}`);
-                    bbIssueUrlCopiedEvent().then(e => { Container.analyticsClient.sendTrackEvent(e); });
+                    bbIssueUrlCopiedEvent().then(e => {
+                        Container.analyticsClient.sendTrackEvent(e);
+                    });
                     break;
                 }
                 case 'assign': {
-                    handled = true;
-                    try {
-                        const bbApi = await clientForSite(this._issue!.site);
-                        await bbApi.issues!.assign(this._issue!, this._issue!.site.details.userId);
-                        await this.update(this._issue!);
-                    } catch (e) {
-                        Logger.error(new Error(`error updating issue: ${e}`));
-                        this.postMessage({ type: 'error', reason: this.formatErrorReason(e) });
+                    if (isAssignIssue(e)) {
+                        handled = true;
+                        try {
+                            const bbApi = await clientForSite(this._issue!.site);
+                            await bbApi.issues!.assign(this._issue!, e.accountId);
+                            await this.update(this._issue!);
+                        } catch (e) {
+                            Logger.error(new Error(`error updating issue: ${e}`));
+                            this.postMessage({ type: 'error', reason: this.formatErrorReason(e) });
+                        }
                     }
-
                     break;
                 }
                 case 'comment': {
@@ -147,12 +156,13 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
                             const bbApi = await clientForSite(this._issue!.site);
                             await bbApi.issues!.postComment(this._issue!, e.content);
                             await this.update(this._issue!);
-                            bbIssueCommentEvent(this._issue!.site.details).then(e => Container.analyticsClient.sendTrackEvent(e));
+                            bbIssueCommentEvent(this._issue!.site.details).then(e =>
+                                Container.analyticsClient.sendTrackEvent(e)
+                            );
                         } catch (e) {
                             Logger.error(new Error(`error posting comment: ${e}`));
                             this.postMessage({ type: 'error', reason: this.formatErrorReason(e) });
                         }
-
                     }
                     break;
                 }
@@ -164,12 +174,13 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
                             await bbApi.issues!.postChange(this._issue!, e.newStatus, e.content);
                             this._issue = await bbApi.issues!.refetch(this._issue!);
                             await this.update(this._issue!);
-                            bbIssueTransitionedEvent(this._issue!.site.details).then(e => Container.analyticsClient.sendTrackEvent(e));
+                            bbIssueTransitionedEvent(this._issue!.site.details).then(e =>
+                                Container.analyticsClient.sendTrackEvent(e)
+                            );
                         } catch (e) {
                             Logger.error(new Error(`error posting change: ${e}`));
                             this.postMessage({ type: 'error', reason: this.formatErrorReason(e) });
                         }
-
                     }
                     break;
                 }
@@ -191,10 +202,8 @@ export class BitbucketIssueWebview extends AbstractReactWebview implements Initi
                     if (isFetchUsers(e)) {
                         handled = true;
                         try {
-                            // TODO: Fix this after pullrequests api uses site instead of remote
-                            // const bbApi = await clientForSite(this._issue!.site);
-                            // const reviewers = await bbApi.pullrequests.getReviewers(e.remote, e.query);
-                            const reviewers = [];
+                            const bbApi = await clientForSite(this._issue!.site);
+                            const reviewers = await bbApi.pullrequests.getReviewers(this._issue!.site, e.query);
                             if (reviewers.length === 0) {
                                 reviewers.push(...this._participants.values());
                             }