+++ src/webviews/components/pullrequest/PullRequestPage.tsx
@@ -22,7 +22,7 @@ import { ApprovalStatus, BitbucketIssue, FileDiff, MergeStrategy, Task } from '.
 import { OpenBitbucketIssueAction } from '../../../ipc/bitbucketIssueActions';
 import { OpenJiraIssueAction } from '../../../ipc/issueActions';
 import { HostErrorMessage, PMFData } from '../../../ipc/messaging';
-import { Checkout, CopyPullRequestLink, CreateTask, DeleteComment, DeleteTask, EditComment, EditTask, FetchUsers, Merge, OpenBuildStatusAction, OpenDiffViewAction, PostComment, RefreshPullRequest, UpdateApproval, UpdateTitle } from '../../../ipc/prActions';
+import { AddReviewer, Checkout, CopyPullRequestLink, CreateTask, DeleteComment, DeleteTask, EditComment, EditTask, FetchUsers, Merge, OpenBuildStatusAction, OpenDiffViewAction, PostComment, RefreshPullRequest, UpdateApproval, UpdateTitle } from '../../../ipc/prActions';
 import { CheckoutResult, isPRData, PRData } from '../../../ipc/prMessaging';
 import { AtlLoader } from '../AtlLoader';
 import BitbucketIssueList from '../bbissue/BitbucketIssueList';
@@ -47,7 +47,7 @@ import Reviewers from './Reviewers';
 
 type Emit = CreateTask |
     EditTask | DeleteTask | UpdateTitle |
-    UpdateApproval | Merge | Checkout |
+    AddReviewer | UpdateApproval | Merge | Checkout |
     PostComment | DeleteComment | EditComment |
     CopyPullRequestLink | OpenJiraIssueAction | OpenBitbucketIssueAction |
     OpenBuildStatusAction | RefreshPullRequest | FetchUsers | OpenDiffViewAction;
@@ -60,9 +60,11 @@ interface ViewState {
     isMergeButtonLoading: boolean;
     isCheckoutButtonLoading: boolean;
     isAnyCommentLoading: boolean;
+    isReviewersLoading: boolean;
     mergeDialogOpen: boolean;
     issueSetupEnabled: boolean;
-    commitMessage: string;
+    commitMessage?: string;
+    defaultCommitMessage: string;
     mergeStrategy: { label: string, value: string | undefined };
     closeSourceBranch?: boolean;
     isErrorBannerOpen: boolean;
@@ -87,9 +89,11 @@ const emptyState: ViewState = {
     isMergeButtonLoading: false,
     isCheckoutButtonLoading: false,
     isAnyCommentLoading: false,
+    isReviewersLoading: false,
     mergeDialogOpen: false,
     issueSetupEnabled: false,
-    commitMessage: '',
+    commitMessage: undefined,
+    defaultCommitMessage: '',
     mergeStrategy: { label: 'Default merge strategy', value: undefined },
     closeSourceBranch: undefined,
     isErrorBannerOpen: false,
@@ -121,7 +125,7 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
         this.postMessage({
             action: 'merge',
             mergeStrategy: this.state.mergeStrategy.value,
-            commitMessage: this.state.commitMessage,
+            commitMessage: this.state.commitMessage || this.state.defaultCommitMessage,
             closeSourceBranch: this.state.closeSourceBranch,
             issue: this.state.issueSetupEnabled ? this.state.pr.mainIssue : undefined
         });
@@ -197,6 +201,14 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
         });
     };
 
+    handleReviewerAdded = (accountId: string) => {
+        this.setState({ isReviewersLoading: true });
+        this.postMessage({
+            action: 'addReviewer',
+            accountId: accountId
+        });
+    };
+
     loadUserOptions = (input: string): Promise<any> => {
         return new Promise(resolve => {
             this.userSuggestions = undefined;
@@ -228,7 +240,7 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
     onMessageReceived(e: any): boolean {
         switch (e.type) {
             case 'error': {
-                this.setState({ isApproveButtonLoading: false, isMergeButtonLoading: false, isCheckoutButtonLoading: false, isAnyCommentLoading: false, isErrorBannerOpen: true, errorDetails: e.reason });
+                this.setState({ isApproveButtonLoading: false, isMergeButtonLoading: false, isCheckoutButtonLoading: false, isAnyCommentLoading: false, isReviewersLoading: false, isErrorBannerOpen: true, errorDetails: e.reason });
                 break;
             }
             case 'checkout': {
@@ -236,6 +248,7 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
                     isApproveButtonLoading: false,
                     isMergeButtonLoading: false,
                     isCheckoutButtonLoading: false,
+                    isReviewersLoading: false,
                     pr: { ...this.state.pr, currentBranch: e.currentBranch }
                 });
                 break;
@@ -253,12 +266,13 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
                         isMergeButtonLoading: false,
                         isCheckoutButtonLoading: false,
                         isAnyCommentLoading: false,
+                        isReviewersLoading: false,
                         closeSourceBranch: this.state.closeSourceBranch === undefined ? e.pr!.data.closeSourceBranch : this.state.closeSourceBranch
                     },
                         () => {
-                            if (this.state.mergeStrategy.value === undefined) {
-                                this.handleMergeStrategyChange(this.state.pr.mergeStrategies.find(strategy => strategy.isDefault === true));
-                            }
+                            this.state.mergeStrategy.value === undefined
+                                ? this.handleMergeStrategyChange(this.state.pr.mergeStrategies.find(strategy => strategy.isDefault === true))
+                                : this.setState({ defaultCommitMessage: this.getDefaultCommitMessage(this.state.mergeStrategy as MergeStrategy) });
                         }
                     );
                 }
@@ -324,16 +338,20 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
 
     handleCommitMessageChange = (text: string) => this.setState({ commitMessage: text });
 
-    handleMergeStrategyChange = (item: any) => this.setState({ mergeStrategy: item }, this.resetCommitMessage);
+    handleMergeStrategyChange = (item: any) => this.setState({
+        mergeStrategy: item,
+        commitMessage: undefined,
+        defaultCommitMessage: this.getDefaultCommitMessage(item)
+    });
 
     handleOpenDiffView = (fileDiff: FileDiff) => {
         this.postMessage({ action: 'openDiffView', fileChange: fileDiff.fileChange! });
     };
 
-    resetCommitMessage = () => {
-        const mergeStrategy = this.state.mergeStrategy.value;
-        if (mergeStrategy === 'fast_forward') {
-            this.setState({ commitMessage: '' });
+    getDefaultCommitMessage = (mergeStrategy: MergeStrategy) => {
+        const mergeStrategyValue = mergeStrategy.value;
+        if (mergeStrategyValue === 'fast_forward') {
+            return '';
         }
 
         const { id, source, title, participants } = this.state.pr.pr!.data;
@@ -343,7 +361,7 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
 
         let defaultCommitMessage = `${branchInfo} ${pullRequestInfo}\n\n${title}`;
 
-        if (mergeStrategy === 'squash') {
+        if (mergeStrategyValue === 'squash') {
             const commits = this.state.pr.commits || [];
             // Minor optimization: if there's exactly 1 commit, and the commit
             // message already matches the pull request title, no need to display the
@@ -365,7 +383,7 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
             defaultCommitMessage += `\n\n${approverInfo}`;
         }
 
-        this.setState({ commitMessage: defaultCommitMessage });
+        return defaultCommitMessage;
     };
 
     render() {
@@ -422,7 +440,7 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
 
         const actionsContent = (
             <div className='ac-inline-grid'>
-                <Reviewers {...this.state.pr} />
+                <Reviewers {...this.state.pr} loadUserOptions={this.loadUserOptions} onAddReviewer={this.handleReviewerAdded} isLoading={this.state.isReviewersLoading} />
                 {!pr.siteDetails.isCloud &&
                     <Tooltip content={currentUserApprovalStatus === 'NEEDS_WORK' ? 'Remove Needs work' : 'Mark as Needs work'}>
                         <Button className={currentUserApprovalStatus === 'NEEDS_WORK' ? 'ac-button-warning' : 'ac-button'}
@@ -451,16 +469,16 @@ export default class PullRequestPage extends WebviewComponent<Emit, Receive, {},
                                         className="ac-select-container"
                                         classNamePrefix="ac-select"
                                         getOptionLabel={(option: MergeStrategy) => `${option.label}${option.isDefault ? ' (default)' : ''}`}
-                                        getOptionValue={(option: MergeStrategy) => option.value}
+                                        getOptionValue={(option: MergeStrategy) => option}
                                         value={this.state.mergeStrategy}
                                         onChange={this.handleMergeStrategyChange} />
                                 </div>
                                 {this.state.mergeStrategy.value !== 'fast_forward' &&
                                     <div className='ac-vpadding'>
-                                        <Tooltip content={this.state.commitMessage}>
+                                        <Tooltip content={this.state.commitMessage || this.state.defaultCommitMessage}>
                                             <EdiText
                                                 type='textarea'
-                                                value={this.state.commitMessage}
+                                                value={this.state.commitMessage || this.state.defaultCommitMessage}
                                                 onSave={this.handleCommitMessageChange}
                                                 validation={isValidString}
                                                 validationMessage='Commit message is required'